+++ Assets/Scripts/World/Tutorial/AnimationsOnDamageGet.cs
@@ -4,7 +4,7 @@ public class AnimationsOnDamageGet : MonoBehaviour, IDamageble
 {
     [SerializeField] Animator animator;
     [SerializeField] string DamageTriggerName;
-    public void OnDamageGet(int Damage,IDamageble.DamageType type)
+    public void OnDamageGet(float Damage,IDamageble.DamageType type)
     {
         print(DamageTriggerName);
         animator.SetTrigger(DamageTriggerName);
