+++ Assets/Scripts/World/Tutorial/Explosive.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 public class Explosive : MonoBehaviour, IDamageble
 {
     [SerializeField] GameObject Explosion;
-    public void OnDamageGet(int Damage,IDamageble.DamageType type)
+    public void OnDamageGet(float Damage,IDamageble.DamageType type)
     {
         Instantiate(Explosion,transform.position,Quaternion.identity);
         Level.EXP += 3;
